- Add test for show logo condition of LoginScreen.kt
